refactor: extract new reward issued modal
